Add files via upload
This has the functionality to check multiple models for high leverage and Cook's D values, remove them, store the new smaller dataset and then check all models with both the original data and deleveraged data for Ridge, Lasso and deliver the results. At this point it seems like overkill, but I figure might as well post the work I did.